requirements.txt
pyyaml==6.0.2
anthropic>=0.40.0
python-dotenv>=1.0.0
moviepy==2.2.1
scipy>=1.14.0
numpy>=1.26.0
Pillow>=10.0.0
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
